Move found repositories file to root folder
The found_repositories.json file was previously stored in the dataset
dir, but it makes sense to only have breaking update json files in this
directory. This move makes it easier to apply JSON schema validation and
prevents the found repo file from being lost among the thousands of
other files.