Make ResultManager create subdirectories for Maven logs
The Maven log files are stored in two subdirectories (successful/unreproducible) in the given reproduction directory. Previously, there was no check to see if these existed, which lead to issues when a directory was provided that didn't have this folder structure. The ResultManager constructor has now been changed to create these directories if they don't already exist.